+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - New: Items in open Display Cases can now be picked using the "Pick Block" key
 - New: The Sentry Remote Access tool now displays the last known name of a Sentry if the Sentry is out of range
+- New: The Portable Radar now always shows its coordinates in messages, even when it has a custom name
 - Change: Sentries now shoot half as frequently to match the shooting rate in newer Minecraft versions
 - Fix: Duplication exploit involving keycards
 - Fix: Duplication exploit involving Reinforced Pistons